Moved milestone and ProjectAttribute.
